services/ml-service/pyproject.toml
[tool.poetry]
name = "fraud-detection-ml-service"
version = "1.0.0"
description = "ML inference microservice — FastAPI REST + gRPC for fraud detection"
authors = ["Fraud Detection Team <team@frauddetection.local>"]

[tool.poetry.dependencies]
python = "^3.11"

# --- Core ML (mirrors root pyproject.toml) ---
numpy         = "^1.26"
pandas        = "^2.2"
scikit-learn  = "^1.4"
xgboost       = "^2.0"
lightgbm      = "^4.3"
imbalanced-learn = "^0.12"
optuna        = "^3.6"
torch         = "^2.2"
torch-geometric = "^2.5"
shap          = "^0.45"
lime          = "^0.2"
networkx      = "^3.3"
scipy         = "^1.13"

# --- API / gRPC ---
fastapi       = "^0.111"
uvicorn       = {extras = ["standard"], version = "^0.29"}
grpcio        = "^1.63"
grpcio-tools  = "^1.63"
grpcio-health-checking = "^1.63"
protobuf      = "^4.25"
pydantic      = "^2.7"
pydantic-settings = "^2.2"

# --- MLflow ---
mlflow        = "^2.13"

# --- Observability ---
structlog     = "^24.1"
opentelemetry-api  = "^1.24"
opentelemetry-sdk  = "^1.24"
opentelemetry-exporter-jaeger = "^1.20"
opentelemetry-instrumentation-fastapi = "^0.45b0"
prometheus-client = "^0.20"

# --- Data ---
pyarrow       = "^16.0"
joblib        = "^1.4"

# --- HTTP ---
httpx         = "^0.27"
python-dotenv = "^1.0"
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
pytest           = "^8.2"
pytest-cov       = "^5.0"
pytest-asyncio   = "^0.23"
httpx            = "^0.27"
black            = "^24.4"
flake8           = "^7.0"
mypy             = "^1.10"
isort            = "^5.13"
faker            = "^25.0"
locust           = "^2.29"

[build-system]
requires      = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length    = 100
target-version = ["py311"]

[tool.isort]
profile    = "black"
line_length = 100

[tool.mypy]
python_version       = "3.11"
strict               = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths    = ["tests"]
asyncio_mode = "auto"
addopts      = "--cov=app --cov-report=term-missing --cov-fail-under=80"
